+++ home/business/src/main/java/vn/finance/home/business/domain/usecase/AvailableBalanceUseCase.kt
@@ -7,9 +7,6 @@ import vn.finance.home.business.domain.model.AvailableBalanceModel
 import vn.finance.home.business.domain.repository.AvailableBalanceRepository
 import javax.inject.Inject
 
-class AvailableBalanceUseCase @Inject constructor(private val repository: AvailableBalanceRepository) :
-    BaseUseCase<Any, ResultModel<AvailableBalanceModel>>() {
-    override fun execute(vararg params: Any?): Flow<ResultModel<AvailableBalanceModel>> {
-        return repository.getAvailableBalance()
-    }
+class AvailableBalanceUseCase @Inject constructor(private val repository: AvailableBalanceRepository) : BaseUseCase<Any, ResultModel<AvailableBalanceModel>>() {
+    override fun execute(vararg params: Any?): Flow<ResultModel<AvailableBalanceModel>> = repository.getAvailableBalance()
 }