+++ home/presentation/src/main/java/finance/home/presentation/di/DomainModule.kt
@@ -14,7 +14,5 @@ class DomainModule {
 
     @Provides
     @ViewModelScoped
-    fun providesAvailableBalanceUseCase(repository: AvailableBalanceRepository): AvailableBalanceUseCase =
-        AvailableBalanceUseCase(repository = repository)
-
+    fun providesAvailableBalanceUseCase(repository: AvailableBalanceRepository): AvailableBalanceUseCase = AvailableBalanceUseCase(repository = repository)
 }